# docs/flatmates-openapi.yaml
openapi: 3.1.0
info:
  title: 360 Flatmates API
  description: |
    Complete API for the 360 Flatmates web platform — a swipe-based flatmate-finding app for India's young professionals.

    ## Authentication
    All endpoints require authentication via Bearer token in the `Authorization` header, unless marked as public.
    Tokens are obtained via Supabase Auth (Phone OTP + Password).

    ## User Modes
    - **room_poster** — Has a flat, wants to rent out a spare room
    - **co_hunter** — Looking for people to flat-search with
    - **open_to_both** — Flexible, open to either

    ## Base URL
    `/api/v1`
  version: 1.0.0
  contact:
    name: 360 Ghar
    url: https://360ghar.com
  license:
    name: Confidential
    url: https://360ghar.com
servers:
  - url: https://api.360ghar.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local Development
...
paths:
...
  /flatmates/bootstrap:
    get:
      tags: [Bootstrap]
      summary: Bootstrap data
      description: Load profile, catalogs, and activity counts after authentication.
      operationId: getFlatmatesBootstrap
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Bootstrap data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatmatesBootstrap'

  # ===================== CATALOGS =====================
  /flatmates/catalogs:
    get:
      tags: [Catalogs]
      summary: Get business catalogs
      description: Server-driven business metadata (modes, timelines, quiz dimensions, vibe tags, report reasons, icebreakers, room types, furnishing options, etc.).
      operationId: getFlatmatesCatalogs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of catalog entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CatalogEntry'

  # ===================== PROFILES =====================
  /flatmates/profile:
    get:
      tags: [Profiles]
      summary: Get my flatmate profile
      description: Fetch the authenticated user's flatmate profile.
      operationId: getMyFlatmatesProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Flatmate profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatmatesProfile'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Profiles]
      summary: Update my flatmate profile
      description: Update the authenticated user's flatmate profile fields.
      operationId: updateMyFlatmatesProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlatmatesProfileUpdate'
      responses:
        '200':
          description: Updated flatmate profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatmatesProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /flatmates/profiles:
    get:
      tags: [Profiles]
      summary: Discover profiles
      description: Fetch discoverable peer profiles for the swipe deck. Applies deal-breaker (non-negotiables) and discovery filtering.
      operationId: discoverProfiles
      security:
        - BearerAuth: []
      parameters:
        - name: city
          in: query
          schema:
            type: string
          description: Filter by city
        - name: budget_min
          in: query
          schema:
            type: integer
            minimum: 0
          description: Minimum budget
        - name: budget_max
          in: query
          schema:
            type: integer
            minimum: 0
          description: Maximum budget
        - name: move_in
          in: query
          schema:
            type: string
            enum: [immediate, this_month, next_month, flexible]
          description: Move-in timeline filter
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of profiles to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Pagination offset
      responses:
        '200':
          description: List of discoverable peer profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlatmatesPeer'

  # ===================== SWIPES =====================
  /flatmates/swipes:
    post:
      tags: [Swipes]
      summary: Record a swipe
      description: |
        Record a swipe action (like, pass, or super-like) on a property or user profile.
        - For property swipes: set `target_type=property` and provide `property_id`
        - For user swipes: set `target_type=user` and provide `target_user_id`
        - Mutual likes automatically create a match and conversation.
        - Super-likes are capped at 3 per day.
      operationId: recordSwipe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwipeRequest'
      responses:
        '200':
          description: Swipe result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwipeResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: Super-like daily limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: Super-like daily limit reached

  # ===================== LIKES =====================
  /flatmates/likes:
    get:
      tags: [Likes]
      summary: Get incoming likes
      description: Fetch profiles that have swiped right on the current user but haven't been answered yet.
      operationId: getIncomingLikes
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of incoming likes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncomingLikeSummary'

  # ===================== MATCHES =====================
  /flatmates/matches:
    get:
      tags: [Matches]
      summary: Get my matches
      description: Fetch all active matches for the current user.
      operationId: getMyMatches
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchSummary'
  /flatmates/matches/{match_id}/unmatch:
    put:
      tags: [Matches]
      summary: Unmatch
      description: Unmatch a specific match, closing the associated conversation.
      operationId: unmatch
      security:
        - BearerAuth: []
      parameters:
        - name: match_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully unmatched
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Match removed successfully
        '404':
          $ref: '#/components/responses/NotFound'

  # ===================== CONVERSATIONS =====================
  /flatmates/conversations:
    get:
      tags: [Conversations]
      summary: List conversations
      description: Fetch all conversations for the current user, ordered by last message timestamp.
      operationId: listConversations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of conversation summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConversationSummary'
  /flatmates/conversations/{conversation_id}:
    get:
      tags: [Conversations]
      summary: Get conversation detail
      description: Fetch full conversation summary including peer info, unread count, and Q&A state.
      operationId: getConversationDetail
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Conversation detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationSummary'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /flatmates/conversations/{conversation_id}/messages:
    get:
      tags: [Conversations]
      summary: Get conversation messages
      description: Fetch messages for a conversation. Automatically marks unread messages as read.
      operationId: getConversationMessages
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
          description: Fetch messages before this message ID (for pagination)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Number of messages to return
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageOut'

    post:
      tags: [Conversations]
      summary: Send a message
      description: Send a text, image, or visit_request message to a conversation.
      operationId: sendMessage
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOut'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /flatmates/conversations/{conversation_id}/mark-read:
    post:
      tags: [Conversations]
      summary: Mark conversation as read
      description: Mark all peer messages in a conversation as read.
      operationId: markConversationRead
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Conversation marked as read
  /flatmates/conversations/{conversation_id}/qna:
    post:
      tags: [Conversations]
      summary: Save Q&A answers
      description: Save guided Q&A answers for a match. Answers are visible to both parties.
      operationId: saveQnAAnswers
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QnAAnswers'
      responses:
        '200':
          description: Q&A answers saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Q&A answers saved successfully

  # ===================== VISITS =====================
...
  /visits/{visit_id}:
    get:
      tags: [Visits]
      summary: Get visit details
      operationId: getVisitDetails
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Visit details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Visits]
      summary: Update visit
      description: Update visit status or scheduled date. Used for confirm/reschedule/cancel actions.
      operationId: updateVisit
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitUpdate'
      responses:
        '200':
          description: Updated visit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
...
  /flatmates/notifications/{notification_id}:
    put:
      tags: [Notifications]
      summary: Mark notification as read
      operationId: markNotificationRead
      security:
        - BearerAuth: []
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlatmatesNotificationUpdate'
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Notification marked as read
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
...
  schemas:
...
    FlatmatesBootstrap:
      type: object
      required: [profile, catalogs, active_listing_count, conversation_count, unread_message_count]
      properties:
        profile:
          $ref: '#/components/schemas/FlatmatesProfile'
        catalogs:
          type: array
          items:
            $ref: '#/components/schemas/CatalogEntry'
        active_listing_count:
          type: integer
        conversation_count:
          type: integer
        unread_message_count:
          type: integer

    # ===================== CATALOG SCHEMAS =====================
    CatalogEntry:
      type: object
      required: [key, version, payload]
      properties:
        key:
          type: string
          description: Catalog key (e.g., flatmate_modes, move_in_timelines, quiz_dimensions, vibe_tags)
        version:
          type: integer
        payload:
          type: object
          additionalProperties: true
          description: Catalog data specific to the key

    # ===================== PROFILE SCHEMAS =====================
...